requirements.txt
flask
azure-cognitiveservices-vision-customvision
python-dotenv
requests